requirements.txt
Flask[async]
telethon
python-dotenv
gunicorn
